+++ app/packages/files/osf-storage-file.ts
@@ -1,4 +1,5 @@
 import FileModel from 'ember-osf-web/models/file';
+import NodeModel from 'ember-osf-web/models/node';
 import File from 'ember-osf-web/packages/files/file';
 import isUnderStorageLimit from 'ember-osf-web/packages/files/file';
 import CurrentUserService from 'ember-osf-web/services/current-user';
@@ -15,5 +16,18 @@ export default class OsfStorageFile extends File {
             return true;
         }
         return false;
+    }
+
+    get isUnderStorageLimit() {
+        const target = this.fileModel.target as unknown as NodeModel;
+        const storageLimitStatus = target.storage.storageLimitStatus;
+        const isPublic = target.public;
+        if (storageLimitStatus === 'OVER_PUBLIC') {
+            return false;
+        }
+        if (isPublic === false && storageLimitStatus === 'OVER_PRIVATE') {
+            return false;
+        }
+        return true;
     }
 }
